requirements.txt
# Core dependencies for tweet automation
apscheduler==3.10.4
pytz==2023.3
requests==2.31.0
python-dotenv==1.0.0

# Twitter API
tweepy==4.14.0

# AI agents for tweet generation
pyautogen==0.2.18

# Google Sheets integration (if using OAuth)
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gspread==5.12.0